Bump to Prometheus 2.11.0
From upstream:

"
We have just cut the final release of v2.11.0. This contains a number of bug fixes, enhancements and a few features, most prominently the option to compress the write-ahead-log to safe disk space. The full changelog can be found in the GitHub release.

Two things to call out:
* write-ahead-log compression can be enabled with the `--storage.tsdb.wal-compression` flag.
* Alertmanager API version to be used by Prometheus is configured using the `api_version` field in the `alertmanagers` config.
"

https://github.com/prometheus/prometheus/releases/tag/v2.11.0